python/envsubst/requirements.txt
envsubst
pyyaml
